PR: Allow null channels in file to not crash setup.
There is an edge case where having `channels` in an environment file
but an empty list, and `channels` missing from the GHA options/keys
returns an unhelpful error message:

`Error: Cannot read property 'join' of null`

I've setup an example here which runs through all the permutations of
`channels` keys being set, missing, and nulled:

https://github.com/Lnaden/ghatesting/runs/1538195898?check_suite_focus=true

This PR attempts to fix that by relaxing the strict equals in `setup.ts`
from

`if (condaConfig["channels"] === "" && channels !== undefined) {`

to

`if (condaConfig["channels"] === "" && channels != null) {`

which should allow `channels != null` to catch both `null` and
`undefined` for `channels`. I also attempted to add an example to the
GitHub Actions tests (Example 9) to test this edge case specifically.

A few comments to this PR:
* I'm not familiar with Typescript and building GHAs, so this might not be correct and I'm happy to make changes and take feedback.
* If you use an env file with an empty `channels` list with the `conda` client to create an environment, it works fine and just defaults to whatever the `conda` is configured to. In my example above, setting no channels in either the file or the GHA keys works fine and pulls from the `default` channel.
* This behavior admittedly is ambiguous, so if you would rather this edge case throw an error, I would like to request a better error message.